Project02-TwoTierInfrastructure/Codes/requirements.txt
flask
flask-cors
pymysql
boto3
mysql-connector-python